backend/requirements_dev.txt
annotated-types==0.6.0
asgiref==3.7.2
black==22.1.0
blinker==1.6.2
blue==0.9.1
Cerberus==1.3.5
certifi==2023.7.22
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
crispy-bootstrap4==2023.1
distlib==0.3.7
Django==4.2.4
django-cors-headers==4.2.0
django-crispy-forms==2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
exceptiongroup==1.1.3
flake8==4.0.1
ghp-import==2.1.0
idna==3.4
iniconfig==2.0.0
isort==5.13.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
mccabe==0.6.1
mergedeep==1.3.4
mkdocs==1.5.2
mypy-extensions==1.0.0
numpy==1.26.1
packaging==23.1
pathspec==0.11.2
pep517==0.13.1
Pillow==10.1.0
pip-api==0.0.30
pipreqs==0.4.13
platformdirs==3.10.0
plette==0.4.4
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.8.0
pydantic==2.5.2
pydantic_core==2.14.5
pyflakes==2.4.0
PyJWT==2.8.0
pyproj==3.6.1
pytest==7.4.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
pyyaml_env_tag==0.1
requests==2.31.0
requirementslib==3.0.0
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
tifffile==2023.9.26
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.7
watchdog==3.0.0
yarg==0.1.9
